feat: implement refresh token to release new token and implement logout it will remove cookie

+++ src/TheGourmet.Api/Controllers/AuthController.cs
@@ -1,3 +1,4 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using TheGourmet.Application.DTOs.Auth;
 using TheGourmet.Application.Interfaces;
@@ -54,12 +55,35 @@ public class AuthController : ControllerBase
     }
 
     // logout endpoint
+    [Authorize] 
     [HttpPost("logout")]
-    public IActionResult Logout()
-    {
+    public async Task<IActionResult> Logout()
+    {   
+        var refreshToken = Request.Cookies["refresh_token"];
+        if (string.IsNullOrWhiteSpace(refreshToken))
+        {
+            return BadRequest(new { Message = "Refresh token is missing" });
+        }
+        var response = await _authService.LogoutAsync(refreshToken);
+
         // Remove cookies
         _cookieService.RemoveAuthCookies();
 
-        return Ok(new { Message = "Logged out successfully" });
+        return Ok(response);
+    }
+
+    // refresh token endpoint
+    [HttpPost("refresh-token")]
+    public async Task<IActionResult> RefreshToken()
+    {
+        var refreshToken = Request.Cookies["refresh_token"];
+        if (string.IsNullOrEmpty(refreshToken))
+        {
+            return Unauthorized(new { Message = "Refresh token is missing" });
+        }
+
+        var response = await _authService.RefreshTokenAsync(refreshToken);
+        _cookieService.SetAuthCookies(response.AccessToken ?? string.Empty, response.RefreshToken ?? string.Empty);
+        return Ok(response);
     }
 }
